+++ spinetoolbox/config.py
@@ -22,8 +22,8 @@ import os
 
 REQUIRED_SPINE_ENGINE_VERSION = "0.8.7"
 REQUIRED_SPINEDB_API_VERSION = "0.9.9"
-PREFERRED_SPINE_ITEMS_VERSION = "0.2.13"
-LATEST_PROJECT_VERSION = 3
+PREFERRED_SPINE_ITEMS_VERSION = "0.3.0"
+LATEST_PROJECT_VERSION = 4
 
 # Invalid characters for directory names
 # NOTE: "." is actually valid in a directory name but this is